+++ angel_serialize_generator/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.6
 * Support for using `abstract` to create immutable model classes.
+* Add support for custom constructor parameters.
 
 # 2.0.5
 * Deserialization now supports un-serialized `DateTime`.